completed edit and delete views for profile and courses.

+++ src/ClassroomManager/AppCore/Interfaces/IRepositoryAsync.cs
@@ -8,7 +8,7 @@ namespace App.Core.Interfaces
 {
     public interface IRepositoryAsync<T> where T : BaseEntity
     {
-        Task<T> GetByIdAsync(int id);
+        Task<T> GetByIdAsync(long id);
         Task<T> GetByUserAsync(string user);
         Task<List<T>> ListAllAsync();
         Task<List<T>> ListAsync(ISpecification<T> spec);
